+++ spec/dummy_app/app/models/custom_version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CustomVersion < PaperTrail::Version
+  has_many :version_associations, class_name: "CustomVersionAssociation", foreign_key: :version_id, dependent: :destroy
+end